python/requirements.txt
# Google OAuth2 JWT Token Generator Dependencies

# Core dependencies for JWT and OAuth2
PyJWT>=2.8.0
requests>=2.31.0
cryptography>=41.0.0

# Optional: For advanced JWT features
python-jose[cryptography]>=3.3.0

# Optional: Google Cloud client libraries (if needed)
# google-auth>=2.23.0
# google-cloud-aiplatform>=1.35.0

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.7.0
# flake8>=6.0.0